Clean up from unused files and comments. Fix the folders and files structure.
