--- apps/meteor/client/views/unreads/components/Toolbox/Toolbox.tsx
@@ -1,68 +1,0 @@
-// import { IMessage, isRoomFederated, IUser } from '@rocket.chat/core-typings';
-// import { MessageToolbox, MessageToolboxItem } from '@rocket.chat/fuselage';
-// import { useUser, useUserSubscription, useSettings, useTranslation } from '@rocket.chat/ui-contexts';
-// import React, { FC, memo, useMemo } from 'react';
-
-// import { MessageAction } from '../../../../../app/ui-utils/client/lib/MessageAction';
-// import { useRoom } from '../../../room/contexts/RoomContext';
-// import { useToolboxContext } from '../../../room/contexts/ToolboxContext';
-// import { useIsSelecting } from '../../../room/contexts/SelectedMessagesContext';
-// import { MessageActionMenu } from './MessageActionMenu';
-
-// export const Toolbox: FC<{ message: IMessage }> = ({ message }) => {
-// 	const t = useTranslation();
-
-// 	const room = useRoom();
-
-// 	const subscription = useUserSubscription(message.rid);
-// 	const settings = useSettings();
-// 	const user = useUser() as IUser;
-// 	const federationContext = isRoomFederated(room) ? 'federated' : '';
-// 	const context = federationContext || 'message';
-
-// 	const mapSettings = useMemo(() => Object.fromEntries(settings.map((setting) => [setting._id, setting.value])), [settings]);
-
-// 	const messageActions = MessageAction.getButtons({ message, room, user, subscription, settings: mapSettings }, context, 'message');
-
-// 	const menuActions = MessageAction.getButtons({ message, room, user, subscription, settings: mapSettings }, context, 'menu');
-
-// 	const toolbox = useToolboxContext();
-
-// 	const isSelecting = useIsSelecting();
-
-// 	if (isSelecting) {
-// 		return null;
-// 	}
-
-// 	return (
-// 		<MessageToolbox>
-// 			{messageActions.map((action) => (
-// 				<MessageToolboxItem
-// 					onClick={(e): void => {
-// 						e.stopPropagation();
-// 						action.action(e, { message, tabbar: toolbox, room });
-// 					}}
-// 					key={action.id}
-// 					icon={action.icon}
-// 					title={t(action.label)}
-// 					data-qa-id={action.label}
-// 					data-qa-type='message-action-menu'
-// 				/>
-// 			))}
-// 			{menuActions.length > 0 && (
-// 				<MessageActionMenu
-// 					options={menuActions.map((action) => ({
-// 						...action,
-// 						action: (e): void => {
-// 							e.stopPropagation();
-// 							action.action(e, { message, tabbar: toolbox, room });
-// 						},
-// 					}))}
-// 					data-qa-type='message-action-menu-options'
-// 				/>
-// 			)}
-// 		</MessageToolbox>
-// 	);
-// };
-
-// export default memo(Toolbox);
