--- apps/meteor/client/views/unreads/components/Toolbox/index.ts
@@ -1,3 +1,0 @@
-// import { ToolboxWrapper } from './ToolboxWrapper';
-
-// export default ToolboxWrapper;
