+++ apps/meteor/client/views/unreads/components/body/EmptyRoomBody.tsx
@@ -2,11 +2,11 @@ import { Box, Icon } from '@rocket.chat/fuselage';
 import React, { FC, memo } from 'react';
 
 const EmptyRoomBody: FC = () => (
-	<Box display='flex' flexDirection='column' alignItems='center' justifyContent='center' width='full' height='70%'>
+	<Box display='flex' flexDirection='column' alignItems='center' justifyContent='center' width='full' height='80%'>
 		<Icon name='check' size='x60' color='green' />
 		<Box marginBlock='x4'>
 			<Box is='h3' color='default' fontScale='h3'>
-				You have no unread messages for now!
+				You have no unread messages at the moment!
 			</Box>
 		</Box>
 	</Box>
