+++ apps/meteor/client/views/unreads/components/header/headerRoomData.ts
@@ -1,0 +1,194 @@
+import { RoomType } from '@rocket.chat/core-typings';
+
+export const headerRoomData = {
+	_id: 'NSDmqTh3jeMNom523',
+	rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
+	u: {
+		_id: '7A69Y5v8iKeccREmw',
+		username: 'Marianna',
+	},
+	_updatedAt: '2022-11-06T00:47:20.245Z',
+	alert: true,
+	fname: 'test',
+	groupMentions: 0,
+	name: 'Unread Messages',
+	open: true,
+	t: 'd' as RoomType,
+	unread: 1,
+	userMentions: 0,
+	lr: '2022-11-06T00:47:20.245Z',
+	tunread: ['ZQeB4SXP4Bh6mxRo5'],
+	ls: '2022-11-06T00:47:20.245Z',
+	ts: '2022-11-06T00:47:20.245Z',
+	messages: [
+		{
+			_id: 'urBuPzrPg6gutcNy2',
+			rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
+			msg: 'Hello!',
+			ts: '2022-11-06T00:47:20.245Z',
+			u: {
+				_id: 'gYHYtdvGqa8iH9zyw',
+				username: 'test_6921002683',
+				name: 'test',
+			},
+			_updatedAt: '2022-11-06T00:47:20.245Z',
+			urls: [],
+			mentions: [],
+			channels: [],
+			md: [
+				{
+					type: 'PARAGRAPH',
+					value: [
+						{
+							type: 'PLAIN_TEXT',
+							value: 'Hello!',
+						},
+					],
+				},
+			],
+		},
+	],
+	threads: [
+		{
+			tmid: 'ZQeB4SXP4Bh6mxRo5',
+			messages: [
+				{
+					_id: 'NiHASbvxC5zfWz82F',
+					rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
+					tmid: 'ZQeB4SXP4Bh6mxRo5',
+					msg: 'vfvf',
+					ts: '2022-11-06T00:47:20.245Z',
+					u: {
+						_id: 'gYHYtdvGqa8iH9zyw',
+						username: 'test_6921002683',
+						name: 'test',
+					},
+					_updatedAt: '2022-11-06T00:47:20.245Z',
+					urls: [],
+					mentions: [],
+					channels: [],
+					md: [
+						{
+							type: 'PARAGRAPH',
+							value: [
+								{
+									type: 'PLAIN_TEXT',
+									value: 'vfvf',
+								},
+							],
+						},
+					],
+				},
+				{
+					_id: 'ut6mfbqt74yRksyTg',
+					rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
+					tmid: 'ZQeB4SXP4Bh6mxRo5',
+					msg: '..',
+					ts: '2022-11-06T00:47:20.245Z',
+					u: {
+						_id: 'gYHYtdvGqa8iH9zyw',
+						username: 'test_6921002683',
+						name: 'test',
+					},
+					_updatedAt: '2022-11-06T00:47:20.245Z',
+					urls: [],
+					mentions: [],
+					channels: [],
+					md: [
+						{
+							type: 'PARAGRAPH',
+							value: [
+								{
+									type: 'PLAIN_TEXT',
+									value: '..',
+								},
+							],
+						},
+					],
+				},
+				{
+					_id: 'o5YQKT4XuM9mYGThm',
+					rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
+					tmid: 'ZQeB4SXP4Bh6mxRo5',
+					msg: '..',
+					ts: '2022-11-06T00:47:20.245Z',
+					u: {
+						_id: 'gYHYtdvGqa8iH9zyw',
+						username: 'test_6921002683',
+						name: 'test',
+					},
+					_updatedAt: '2022-11-06T00:47:20.245Z',
+					urls: [],
+					mentions: [],
+					channels: [],
+					md: [
+						{
+							type: 'PARAGRAPH',
+							value: [
+								{
+									type: 'PLAIN_TEXT',
+									value: '..',
+								},
+							],
+						},
+					],
+				},
+				{
+					_id: 'eQJj9sYCcW5z3mXZ2',
+					rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
+					tmid: 'ZQeB4SXP4Bh6mxRo5',
+					msg: '??',
+					ts: '2022-11-06T00:47:20.245Z',
+					u: {
+						_id: 'gYHYtdvGqa8iH9zyw',
+						username: 'test_6921002683',
+						name: 'test',
+					},
+					_updatedAt: '2022-11-06T00:47:20.245Z',
+					urls: [],
+					mentions: [],
+					channels: [],
+					md: [
+						{
+							type: 'PARAGRAPH',
+							value: [
+								{
+									type: 'PLAIN_TEXT',
+									value: '??',
+								},
+							],
+						},
+					],
+				},
+				{
+					_id: 'yZvNFk3RiEkbTNEgy',
+					rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
+					tshow: true,
+					tmid: 'ZQeB4SXP4Bh6mxRo5',
+					msg: 'test',
+					ts: '2022-11-06T00:47:20.245Z',
+					u: {
+						_id: 'gYHYtdvGqa8iH9zyw',
+						username: 'test_6921002683',
+						name: 'test',
+					},
+					_updatedAt: '2022-11-06T00:47:20.245Z',
+					urls: [],
+					mentions: [],
+					channels: [],
+					md: [
+						{
+							type: 'PARAGRAPH',
+							value: [
+								{
+									type: 'PLAIN_TEXT',
+									value: 'test',
+								},
+							],
+						},
+					],
+				},
+			],
+		},
+	],
+};
