--- apps/meteor/client/views/unreads/mockRoom.tsx
@@ -1,1259 +1,0 @@
-export const mockRoom = [
-	{
-		_id: 'NSDmqTh3jeMNom523',
-		rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-		u: {
-			_id: '7A69Y5v8iKeccREmw',
-			username: 'Marianna',
-		},
-		_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-		alert: true,
-		fname: 'test',
-		groupMentions: 0,
-		name: 'Marianna',
-		open: true,
-		t: 'd',
-		unread: 1,
-		userMentions: 0,
-		lr: new Date('2022-11-06T00:47:20.245Z'),
-		tunread: ['ZQeB4SXP4Bh6mxRo5'],
-		ls: new Date('2022-11-06T00:47:20.245Z'),
-		ts: new Date('2022-11-06T00:47:20.245Z'),
-		messages: [
-			{
-				_id: 'urBuPzrPg6gutcNy2',
-				rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-				msg: 'Hello!',
-				ts: new Date('2022-11-06T00:47:20.245Z'),
-				u: {
-					_id: 'gYHYtdvGqa8iH9zyw',
-					username: 'test_6921002683',
-					name: 'test',
-				},
-				_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-				urls: [],
-				mentions: [],
-				channels: [],
-				md: [
-					{
-						type: 'PARAGRAPH',
-						value: [
-							{
-								type: 'PLAIN_TEXT',
-								value: 'Hello!',
-							},
-						],
-					},
-				],
-			},
-			{
-				_id: 'urBuPzrPg6gutcNy2',
-				rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-				msg: 'Aloha!',
-				ts: new Date('2022-11-06T00:47:20.245Z'),
-				u: {
-					_id: 'gYHYtdvGqa8iH9zyw',
-					username: 'test_6921002683',
-					name: 'test',
-				},
-				_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-				urls: [],
-				mentions: [],
-				channels: [],
-				md: [
-					{
-						type: 'PARAGRAPH',
-						value: [
-							{
-								type: 'PLAIN_TEXT',
-								value: 'Aloha!',
-							},
-						],
-					},
-				],
-			},
-			{
-				_id: 'urBuPzrPg6gutcNy2',
-				rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-				msg: 'Yahoooooo...!',
-				ts: new Date('2022-11-06T00:47:20.245Z'),
-				u: {
-					_id: 'gYHYtdvGqa8iH9zyw',
-					username: 'test_6921002683',
-					name: 'test',
-				},
-				_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-				urls: [],
-				mentions: [],
-				channels: [],
-				md: [
-					{
-						type: 'PARAGRAPH',
-						value: [
-							{
-								type: 'PLAIN_TEXT',
-								value: 'Yahoooooo...!',
-							},
-						],
-					},
-				],
-			},
-			{
-				_id: 'urBuPzrPg6gutcNy2',
-				rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-				msg: 'Ti kanete???',
-				ts: new Date('2022-11-06T00:47:20.245Z'),
-				u: {
-					_id: 'gYHYtdvGqa8iH9zyw',
-					username: 'test_6921002683',
-					name: 'test',
-				},
-				_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-				urls: [],
-				mentions: [],
-				channels: [],
-				md: [
-					{
-						type: 'PARAGRAPH',
-						value: [
-							{
-								type: 'PLAIN_TEXT',
-								value: 'Ti kanete???',
-							},
-						],
-					},
-				],
-			},
-			{
-				_id: 'urBuPzrPg6gutcNy2',
-				rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-				msg: 'Opaaaaaa!',
-				ts: new Date('2022-11-06T00:47:20.245Z'),
-				u: {
-					_id: 'gYHYtdvGqa8iH9zyw',
-					username: 'test_6921002683',
-					name: 'test',
-				},
-				_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-				urls: [],
-				mentions: [],
-				channels: [],
-				md: [
-					{
-						type: 'PARAGRAPH',
-						value: [
-							{
-								type: 'PLAIN_TEXT',
-								value: 'Opaaaaaa!',
-							},
-						],
-					},
-				],
-			},
-		],
-		threads: [
-			{
-				tmid: 'ZQeB4SXP4Bh6mxRo5',
-				messages: [
-					{
-						_id: 'NiHASbvxC5zfWz82F',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: 'vfvf',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: 'vfvf',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'ut6mfbqt74yRksyTg',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: '..',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: '..',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'o5YQKT4XuM9mYGThm',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: '..',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: '..',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'eQJj9sYCcW5z3mXZ2',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: '??',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: '??',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'yZvNFk3RiEkbTNEgy',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tshow: true,
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: 'test',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: 'test',
-									},
-								],
-							},
-						],
-					},
-				],
-			},
-		],
-	},
-	{
-		_id: 'NSDmqTh3jeMNom523',
-		rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-		u: {
-			_id: '7A69Y5v8iKeccREmw',
-			username: 'Marianna',
-		},
-		_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-		alert: true,
-		fname: 'test',
-		groupMentions: 0,
-		name: 'Lampros',
-		open: true,
-		t: 'd',
-		unread: 1,
-		userMentions: 0,
-		lr: new Date('2022-11-06T00:47:20.245Z'),
-		tunread: ['ZQeB4SXP4Bh6mxRo5'],
-		ls: new Date('2022-11-06T00:47:20.245Z'),
-		ts: new Date('2022-11-06T00:47:20.245Z'),
-		messages: [
-			{
-				_id: 'urBuPzrPg6gutcNy2',
-				rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-				msg: 'Hello!',
-				ts: new Date('2022-11-06T00:47:20.245Z'),
-				u: {
-					_id: 'gYHYtdvGqa8iH9zyw',
-					username: 'test_6921002683',
-					name: 'test',
-				},
-				_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-				urls: [],
-				mentions: [],
-				channels: [],
-				md: [
-					{
-						type: 'PARAGRAPH',
-						value: [
-							{
-								type: 'PLAIN_TEXT',
-								value: 'Hello!',
-							},
-						],
-					},
-				],
-			},
-		],
-		threads: [
-			{
-				tmid: 'ZQeB4SXP4Bh6mxRo5',
-				messages: [
-					{
-						_id: 'NiHASbvxC5zfWz82F',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: 'vfvf',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: 'vfvf',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'ut6mfbqt74yRksyTg',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: '..',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: '..',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'o5YQKT4XuM9mYGThm',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: '..',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: '..',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'eQJj9sYCcW5z3mXZ2',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: '??',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: '??',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'yZvNFk3RiEkbTNEgy',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tshow: true,
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: 'test',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: 'test',
-									},
-								],
-							},
-						],
-					},
-				],
-			},
-		],
-	},
-	{
-		_id: 'NSDmqTh3jeMNom523',
-		rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-		u: {
-			_id: '7A69Y5v8iKeccREmw',
-			username: 'Marianna',
-		},
-		_updatedAt: '2022-11-06T00:47:20.245Z',
-		alert: true,
-		fname: 'test',
-		groupMentions: 0,
-		name: 'Vasilis',
-		open: true,
-		t: 'd',
-		unread: 1,
-		userMentions: 0,
-		lr: '2022-11-06T00:47:20.245Z',
-		tunread: ['ZQeB4SXP4Bh6mxRo5'],
-		ls: '2022-11-06T00:47:20.245Z',
-		ts: '2022-11-06T00:47:20.245Z',
-		messages: [
-			{
-				_id: 'urBuPzrPg6gutcNy2',
-				rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-				msg: 'Hello!',
-				ts: '2022-11-06T00:47:20.245Z',
-				u: {
-					_id: 'gYHYtdvGqa8iH9zyw',
-					username: 'test_6921002683',
-					name: 'test',
-				},
-				_updatedAt: '2022-11-06T00:47:20.245Z',
-				urls: [],
-				mentions: [],
-				channels: [],
-				md: [
-					{
-						type: 'PARAGRAPH',
-						value: [
-							{
-								type: 'PLAIN_TEXT',
-								value: 'Hello!',
-							},
-						],
-					},
-				],
-			},
-		],
-		threads: [
-			{
-				tmid: 'ZQeB4SXP4Bh6mxRo5',
-				messages: [
-					{
-						_id: 'NiHASbvxC5zfWz82F',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: 'vfvf',
-						ts: '2022-11-06T00:47:20.245Z',
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: '2022-11-06T00:47:20.245Z',
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: 'vfvf',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'ut6mfbqt74yRksyTg',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: '..',
-						ts: '2022-11-06T00:47:20.245Z',
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: '2022-11-06T00:47:20.245Z',
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: '..',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'o5YQKT4XuM9mYGThm',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: '..',
-						ts: '2022-11-06T00:47:20.245Z',
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: '2022-11-06T00:47:20.245Z',
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: '..',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'eQJj9sYCcW5z3mXZ2',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: '??',
-						ts: '2022-11-06T00:47:20.245Z',
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: '2022-11-06T00:47:20.245Z',
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: '??',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'yZvNFk3RiEkbTNEgy',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tshow: true,
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: 'test',
-						ts: '2022-11-06T00:47:20.245Z',
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: '2022-11-06T00:47:20.245Z',
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: 'test',
-									},
-								],
-							},
-						],
-					},
-				],
-			},
-		],
-	},
-	{
-		_id: 'NSDmqTh3jeMNom523',
-		rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-		u: {
-			_id: '7A69Y5v8iKeccREmw',
-			username: 'Marianna',
-		},
-		_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-		alert: true,
-		fname: 'test',
-		groupMentions: 0,
-		name: 'Unread Messages',
-		open: true,
-		t: 'd',
-		unread: 1,
-		userMentions: 0,
-		lr: new Date('2022-11-06T00:47:20.245Z'),
-		tunread: ['ZQeB4SXP4Bh6mxRo5'],
-		ls: new Date('2022-11-06T00:47:20.245Z'),
-		ts: new Date('2022-11-06T00:47:20.245Z'),
-		messages: [
-			{
-				_id: 'urBuPzrPg6gutcNy2',
-				rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-				msg: 'Hello!',
-				ts: new Date('2022-11-06T00:47:20.245Z'),
-				u: {
-					_id: 'gYHYtdvGqa8iH9zyw',
-					username: 'test_6921002683',
-					name: 'test',
-				},
-				_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-				urls: [],
-				mentions: [],
-				channels: [],
-				md: [
-					{
-						type: 'PARAGRAPH',
-						value: [
-							{
-								type: 'PLAIN_TEXT',
-								value: 'Hello!',
-							},
-						],
-					},
-				],
-			},
-		],
-		threads: [
-			{
-				tmid: 'ZQeB4SXP4Bh6mxRo5',
-				messages: [
-					{
-						_id: 'NiHASbvxC5zfWz82F',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: 'vfvf',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: 'vfvf',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'ut6mfbqt74yRksyTg',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: '..',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: '..',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'o5YQKT4XuM9mYGThm',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: '..',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: '..',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'eQJj9sYCcW5z3mXZ2',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: '??',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: '??',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'yZvNFk3RiEkbTNEgy',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tshow: true,
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: 'test',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: 'test',
-									},
-								],
-							},
-						],
-					},
-				],
-			},
-		],
-	},
-	{
-		_id: 'NSDmqTh3jeMNom523',
-		rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-		u: {
-			_id: '7A69Y5v8iKeccREmw',
-			username: 'Marianna',
-		},
-		_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-		alert: true,
-		fname: 'test',
-		groupMentions: 0,
-		name: 'test_6921002683',
-		open: true,
-		t: 'd',
-		unread: 1,
-		userMentions: 0,
-		lr: new Date('2022-11-06T00:47:20.245Z'),
-		tunread: ['ZQeB4SXP4Bh6mxRo5'],
-		ls: new Date('2022-11-06T00:47:20.245Z'),
-		ts: new Date('2022-11-06T00:47:20.245Z'),
-		messages: [
-			{
-				_id: 'urBuPzrPg6gutcNy2',
-				rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-				msg: 'Hello!',
-				ts: new Date('2022-11-06T00:47:20.245Z'),
-				u: {
-					_id: 'gYHYtdvGqa8iH9zyw',
-					username: 'test_6921002683',
-					name: 'test',
-				},
-				_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-				urls: [],
-				mentions: [],
-				channels: [],
-				md: [
-					{
-						type: 'PARAGRAPH',
-						value: [
-							{
-								type: 'PLAIN_TEXT',
-								value: 'Hello!',
-							},
-						],
-					},
-				],
-			},
-		],
-		threads: [
-			{
-				tmid: 'ZQeB4SXP4Bh6mxRo5',
-				messages: [
-					{
-						_id: 'NiHASbvxC5zfWz82F',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: 'vfvf',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: 'vfvf',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'ut6mfbqt74yRksyTg',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: '..',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: '..',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'o5YQKT4XuM9mYGThm',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: '..',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: '..',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'eQJj9sYCcW5z3mXZ2',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: '??',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: '??',
-									},
-								],
-							},
-						],
-					},
-					{
-						_id: 'yZvNFk3RiEkbTNEgy',
-						rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-						tshow: true,
-						tmid: 'ZQeB4SXP4Bh6mxRo5',
-						msg: 'test',
-						ts: new Date('2022-11-06T00:47:20.245Z'),
-						u: {
-							_id: 'gYHYtdvGqa8iH9zyw',
-							username: 'test_6921002683',
-							name: 'test',
-						},
-						_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-						urls: [],
-						mentions: [],
-						channels: [],
-						md: [
-							{
-								type: 'PARAGRAPH',
-								value: [
-									{
-										type: 'PLAIN_TEXT',
-										value: 'test',
-									},
-								],
-							},
-						],
-					},
-				],
-			},
-		],
-	},
-];
-
-export const headerRoomData = {
-	_id: 'NSDmqTh3jeMNom523',
-	rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-	u: {
-		_id: '7A69Y5v8iKeccREmw',
-		username: 'Marianna',
-	},
-	_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-	alert: true,
-	fname: 'test',
-	groupMentions: 0,
-	name: 'Unread Messages',
-	open: true,
-	t: 'd',
-	unread: 1,
-	userMentions: 0,
-	lr: new Date('2022-11-06T00:47:20.245Z'),
-	tunread: ['ZQeB4SXP4Bh6mxRo5'],
-	ls: new Date('2022-11-06T00:47:20.245Z'),
-	ts: new Date('2022-11-06T00:47:20.245Z'),
-	messages: [
-		{
-			_id: 'urBuPzrPg6gutcNy2',
-			rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-			msg: 'Hello!',
-			ts: new Date('2022-11-06T00:47:20.245Z'),
-			u: {
-				_id: 'gYHYtdvGqa8iH9zyw',
-				username: 'test_6921002683',
-				name: 'test',
-			},
-			_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-			urls: [],
-			mentions: [],
-			channels: [],
-			md: [
-				{
-					type: 'PARAGRAPH',
-					value: [
-						{
-							type: 'PLAIN_TEXT',
-							value: 'Hello!',
-						},
-					],
-				},
-			],
-		},
-	],
-	threads: [
-		{
-			tmid: 'ZQeB4SXP4Bh6mxRo5',
-			messages: [
-				{
-					_id: 'NiHASbvxC5zfWz82F',
-					rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-					tmid: 'ZQeB4SXP4Bh6mxRo5',
-					msg: 'vfvf',
-					ts: new Date('2022-11-06T00:47:20.245Z'),
-					u: {
-						_id: 'gYHYtdvGqa8iH9zyw',
-						username: 'test_6921002683',
-						name: 'test',
-					},
-					_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-					urls: [],
-					mentions: [],
-					channels: [],
-					md: [
-						{
-							type: 'PARAGRAPH',
-							value: [
-								{
-									type: 'PLAIN_TEXT',
-									value: 'vfvf',
-								},
-							],
-						},
-					],
-				},
-				{
-					_id: 'ut6mfbqt74yRksyTg',
-					rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-					tmid: 'ZQeB4SXP4Bh6mxRo5',
-					msg: '..',
-					ts: new Date('2022-11-06T00:47:20.245Z'),
-					u: {
-						_id: 'gYHYtdvGqa8iH9zyw',
-						username: 'test_6921002683',
-						name: 'test',
-					},
-					_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-					urls: [],
-					mentions: [],
-					channels: [],
-					md: [
-						{
-							type: 'PARAGRAPH',
-							value: [
-								{
-									type: 'PLAIN_TEXT',
-									value: '..',
-								},
-							],
-						},
-					],
-				},
-				{
-					_id: 'o5YQKT4XuM9mYGThm',
-					rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-					tmid: 'ZQeB4SXP4Bh6mxRo5',
-					msg: '..',
-					ts: new Date('2022-11-06T00:47:20.245Z'),
-					u: {
-						_id: 'gYHYtdvGqa8iH9zyw',
-						username: 'test_6921002683',
-						name: 'test',
-					},
-					_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-					urls: [],
-					mentions: [],
-					channels: [],
-					md: [
-						{
-							type: 'PARAGRAPH',
-							value: [
-								{
-									type: 'PLAIN_TEXT',
-									value: '..',
-								},
-							],
-						},
-					],
-				},
-				{
-					_id: 'eQJj9sYCcW5z3mXZ2',
-					rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-					tmid: 'ZQeB4SXP4Bh6mxRo5',
-					msg: '??',
-					ts: new Date('2022-11-06T00:47:20.245Z'),
-					u: {
-						_id: 'gYHYtdvGqa8iH9zyw',
-						username: 'test_6921002683',
-						name: 'test',
-					},
-					_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-					urls: [],
-					mentions: [],
-					channels: [],
-					md: [
-						{
-							type: 'PARAGRAPH',
-							value: [
-								{
-									type: 'PLAIN_TEXT',
-									value: '??',
-								},
-							],
-						},
-					],
-				},
-				{
-					_id: 'yZvNFk3RiEkbTNEgy',
-					rid: '7A69Y5v8iKeccREmwgYHYtdvGqa8iH9zyw',
-					tshow: true,
-					tmid: 'ZQeB4SXP4Bh6mxRo5',
-					msg: 'test',
-					ts: new Date('2022-11-06T00:47:20.245Z'),
-					u: {
-						_id: 'gYHYtdvGqa8iH9zyw',
-						username: 'test_6921002683',
-						name: 'test',
-					},
-					_updatedAt: new Date('2022-11-06T00:47:20.245Z'),
-					urls: [],
-					mentions: [],
-					channels: [],
-					md: [
-						{
-							type: 'PARAGRAPH',
-							value: [
-								{
-									type: 'PLAIN_TEXT',
-									value: 'test',
-								},
-							],
-						},
-					],
-				},
-			],
-		},
-	],
-};
